requirements.txt
django
Pillow
celery==4.4.2
flower
braintree
WeasyPrint
python-dotenv
uwsgi==2.0.18
PyYAML==5.3.1
python-dotenv
psycopg2
